Orders/requirements.txt
django
djangorestframework
grpcio
grpcio-tools
djangorestframework-simplejwt
python-dotenv
psycopg
confluent_kafka